+++ src/commands/mod.rs
@@ -20,7 +20,6 @@ pub mod history;
 pub mod keywords;
 pub mod launch;
 pub mod list_devices;
-pub mod local_models;
 pub mod logs;
 pub mod model;
 pub mod process;
@@ -35,7 +34,6 @@ pub use history::handle_history;
 pub use keywords::handle_keywords;
 pub use launch::handle_launch;
 pub use list_devices::handle_list_devices;
-pub use local_models::handle_local_models;
 pub use logs::handle_logs;
 pub use model::handle_model;
 pub use process::handle_process;
