+++ src/transcription/api/groq.rs
@@ -19,7 +19,7 @@ struct GroqResponse {
 /// Groq provides an OpenAI-compatible API endpoint.
 ///
 /// Keywords are passed as the `prompt` parameter to guide transcription context.
-pub async fn transcribe(config: &TranscriptionConfig, audio_path: &Path) -> anyhow::Result<String> {
+pub(super) async fn transcribe(config: &TranscriptionConfig, audio_path: &Path) -> anyhow::Result<String> {
     let audio_data =
         std::fs::read(audio_path).map_err(|e| anyhow::anyhow!("Failed to read audio file: {e}"))?;
